feat(CC): Add creation timestamp to accounts
Introduces a `createdAt` field to the `CCAccountData` interface in `types.ts` to store the creation timestamp of current accounts. This change also updates `CurrentAccountsApp.tsx` to initialize this field when a new account is created.

Additionally, `MenuModal.tsx` now accepts a Firebase database reference (`db`) and includes logic to prepare for downloading data, including account creation timestamps, to facilitate backups. `App.tsx` is updated to pass the `db` object to the `MenuModal`.